Updated image name, updated workflow steps actions, fixed Docker image
Hopefully the last rename. adding container to every image name is way too verbose, but it is nice to have for the repository name.
Solution is to just strip it for the container image registries, so we will do this from now on.

Docker image was also still pushing to vintage-story-server.. whoops.
Also, as of this commit, the 1-year old deprecated images will be deleted.